Make new changes to still-slider.css
Made a couple of styling changes re #imageDescription which became #quiz_description so that it would align with all other markup related to the quiz. Made the corresponding changes in still-slider.js as well. Also, hid the slider buttons. Didn't want to remove them. As for the quiz_description heading that the player hovers over to reveal the quiz_description text, I borrowed from the navigation menu markup. So I had to add it to media screen and min-width: 832px as well as above 1025px so that it would appear. The original navigation markup hides it at those widths.